backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
supabase>=2.0.3
pydantic>=2.5.0
python-dateutil>=2.8.2
google-generativeai>=0.3.2
httpx>=0.25.2
